backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
torch>=2.6.0
torchvision>=0.21.0
Pillow>=10.3.0
numpy>=1.26.4
opencv-python-headless>=4.9.0.80
fpdf2>=2.7.9
